VMM: More warnings.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@26279 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMAll/PGMAllPool.cpp
@@ -966,7 +966,7 @@ DECLINLINE(int) pgmPoolAccessHandlerSTOSD(PVM pVM, PPGMPOOL pPool, PPGMPOOLPAGE 
         pgmPoolMonitorChainChanging(pVCpu, pPool, pPage, GCPhysFault, (RTGCPTR)pu32, uIncrement);
 #endif
 #ifdef IN_RC
-        *(uint32_t *)pu32 = pRegFrame->eax;
+        *(uint32_t *)(uintptr_t)pu32 = pRegFrame->eax;
 #else
         PGMPhysSimpleWriteGCPhys(pVM, GCPhysFault, &pRegFrame->rax, uIncrement);
 #endif
